Refactor how results are returned to the client
This refactors the internal result returning system to match more
closely how we iterative return points between result sets, series, and
rows within the series. It uses the updated terminology rather than
older terminology that no longer matches how we refer to things in the
documentation or within the query engine.

The refactor moves the aggregation and chunking behavior from
`influxql.Emitter` to the HTTP service so that behavior is isolated to
one location rather than sprinkled around in multiple locations. This
refactor also allows us to use different output structures. One of these
new output structures is implemented when using `application/x-msgpack`
as the output format.

This new format is more suited to client designs than the old JSON
format for a few reasons. First, it maintains floats and integers as
separate types. So, unlike JSON, you do not have to guess if a returned
point is a float or an integer when it is a whole number. That is
returned as part of the format.

Additionally, it removes the current difficult to use chunking and
creates a new iterative format that shouldn't be as difficult to
process. For this new format, there is no non-chunking format.